nested loops

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,17 +8,17 @@
 
 int main(void)
 {
-unsigned int fib1 = 1;
-unsigned int fib2 = 2;
-unsigned int fib_next;
+unsigned long fib1 = 1;
+unsigned long fib2 = 2;
+unsigned long fib_next;
 int count;
 
-printf("%u, %u", fib1, fib2);
+printf("%lu, %lu", fib1, fib2);
 
 for (count = 3; count <= 98; count++)
 {
 fib_next = fib1 + fib2;
-printf(", %u", fib_next);
+printf(", %lu", fib_next);
 
 fib1 = fib2;
 fib2 = fib_next;
